Post-PR refactoring
