+++ pom.xml
@@ -60,7 +60,7 @@
             <url>https://repo.destroystokyo.com/repository/maven-public/</url>
         </repository>
         <repository>
-            <id>worldedit-worldguard-repo</id>
+            <id>worldedit-repo</id>
             <url>https://maven.sk89q.com/repo/</url>
         </repository>
         <repository>
