+++ src/main/java/io/github/thebusybiscuit/slimefun4/implementation/setup/PostSetup.java
@@ -217,9 +217,13 @@ public final class PostSetup {
             }
 
             for (SlimefunItem item : SlimefunPlugin.getRegistry().getEnabledSlimefunItems()) {
-                if (item instanceof AContainer && ((AContainer) item).getMachineIdentifier().equals("ELECTRIC_SMELTERY")) {
-                    List<MachineRecipe> recipes = ((AContainer) item).getMachineRecipes();
-                    Collections.sort(recipes, Comparator.comparingInt(recipe -> recipe == null ? 0 : -recipe.getInput().length));
+                if (item instanceof AContainer) {
+                    AContainer machine = (AContainer) item;
+
+                    if (machine.getMachineIdentifier().equals("ELECTRIC_SMELTERY")) {
+                        List<MachineRecipe> recipes = machine.getMachineRecipes();
+                        Collections.sort(recipes, Comparator.comparingInt(recipe -> recipe == null ? 0 : -recipe.getInput().length));
+                    }
                 }
             }
         }
@@ -254,8 +258,12 @@ public final class PostSetup {
 
     private static void registerMachineRecipe(String machine, int seconds, ItemStack[] input, ItemStack[] output) {
         for (SlimefunItem item : SlimefunPlugin.getRegistry().getEnabledSlimefunItems()) {
-            if (item instanceof AContainer && ((AContainer) item).getMachineIdentifier().equals(machine)) {
-                ((AContainer) item).registerRecipe(seconds, input, output);
+            if (item instanceof AContainer) {
+                AContainer container = (AContainer) item;
+
+                if (container.getMachineIdentifier().equals(machine)) {
+                    container.registerRecipe(seconds, input, output);
+                }
             }
         }
     }
